Fixes done for Improvement in Coverage
Added updates for improving the coverage.
Handled changes in catalog
Added warning message for Birthcertificate.
Added Fix in Comparator for Value Validation.
Added a check to see if fdr service is enabled.
Tool warns if imp file or dir are not present.
Updated Tool Version to 2.2 in fdrtool & setup.py.